+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.64.0] - 2023-02-24
+
 ### Added
 - Shipping option header to `productSearch` and `facets` queries.